Refactor swt_getFunctionWithName() a bit, don't need Allocator.h anymore
