+++ Sources/TestingInternals/Discovery.cpp
@@ -23,8 +23,6 @@
 #include <mach-o/getsect.h>
 #include <objc/runtime.h>
 #include <os/lock.h>
-
-#include "Allocator.h"
 #endif
 
 /// Enumerate over all Swift type metadata sections in the current process.
@@ -178,6 +176,27 @@ static void enumerateTypeMetadataSections(const SectionEnumerator& body) {
 
 #elif defined(__APPLE__)
 #pragma mark - Apple implementation
+
+/// A type that acts as a C++ [Allocator](https://en.cppreference.com/w/cpp/named_req/Allocator)
+/// without using global `operator new` or `operator delete`.
+///
+/// This type is necessary because global `operator new` and `operator delete`
+/// can be overridden in developer-supplied code and cause deadlocks or crashes
+/// when subsequently used while holding a dyld- or libobjc-owned lock. Using
+/// `std::malloc()` and `std::free()` allows the use of C++ container types
+/// without this risk.
+template<typename T>
+struct SWTHeapAllocator {
+  using value_type = T;
+
+  T *allocate(size_t count) {
+    return reinterpret_cast<T *>(std::calloc(count, sizeof(T)));
+  }
+
+  void deallocate(T *ptr, size_t count) {
+    std::free(ptr);
+  }
+};
 
 /// A type that acts as a C++ [Container](https://en.cppreference.com/w/cpp/named_req/Container)
 /// and which contains a sequence of Mach headers.
